+++ js/main.js
@@ -1,5 +1,6 @@
 import {getMockAds} from'./mock-data.js';
 import {drawOffers} from './card-popup.js';
 import './form.js';
+import './map.js';
 
 drawOffers(getMockAds(2));
